option,helm: Define agent and Helm flags for SRv6 support
This agent flag will be used in follow up commits. For now, it simply
results in a new, unused ENABLE_SRV6 macro in the datapath config files.

The operator flag is required because it will condition the CRD
registration (when the operator is in charge of that as in the dev. VM).

As should be obvious from the name, SRv6 requires IPv6 to be enabled in
Cilium.

Signed-off-by: Paul Chaignon <paul@cilium.io>