bpf: Perform SRv6 policy lookup
To perform the SRv6 policy lookup, we must first lookup the appropriate
VRF where to perform the policy lookup. This first lookup is performed
in LPM maps using source and destination addresses. The second lookup
then uses the retrieved VRF ID and the destination address in another
LPM map to get the SID, used as the outer destination IPv6 address.

The lookups should only be performed for connections egressing the
cluster (i.e., same as for the egress gateway). Otherwise, a destination
pod IP address could match an entry in the LPM map if e.g. the user
defined a redirect for destination CIDR 0.0.0.0/0. In addition, we don't
want to perform the lookup and encapsulation for reply traffic (which
should leave the cluster the same way it ingressed).

Signed-off-by: Paul Chaignon <paul@cilium.io>