egressmap: add support for multiple gateway nodes per policy
This commit lays the groundwork (mostly from a datapath perspective) to
implement high availability egress gateway by adding support for
multiple gateway nodes per policy, sharing all the same egress IP.

The full CFP can be found at [1].

The format of the egress policy map has been updated to allow storing
multiple gateways per policy.

Note that these changes assume the router in front of the gateways will
keep track of which ECMP route a connection belongs to (in order to
forward reply traffic back to the correct gateway).
This in practice is not always true, and in a subsequent PR a more
general approach will be introduced to make sure reply traffic will
eventually hit the correct gateway, regardless of which one the router
forwards the traffic to.

In addition to that, this commit introduces also a new map,
cilium_egress_ct_v4, which is used to pin connections to a specific
egress gateway node.

In this way, after a gateway node is randomly picked for a new
connection, all traffic belonging to that connection will be redirected
to the same gateway.

Lastly, for the case of an egress policy with a single gateway, this
change should be backward compatible with the current egress gateway
implementation.

[1] https://docs.google.com/document/d/1ojU8bmy6ai8fdClHCVAPN5-SExm5KjLsh_IJ24tY-4M/edit?usp=sharing

Enterprise-feature: Egress Gateway HA
Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>