k8s: add EgressGroups property to CiliumEgressNATPolicy object
This commit extends the CiliumEgressNATPolicy k8s object to support an
additional field, EgressGroup. This field allows to select a group of
nodes that will act as egress gateway for the policy (and so for the
same egress IP) in a HA setup.

The EgressGroup fields allows to specify:

* the node group label, used to match nodes belonging to the group
* the name of network interface or the egress IP used to SNAT traffic

An example of a HA policy is:

    apiVersion: cilium.io/v2alpha1
    kind: CiliumEgressNATPolicy
    metadata:
      name: egress-policy-name
    spec:
      egress:
      - podSelector:
	  matchLabels:
	    a: b
      destinationCIDRs:
      - 1.1.1.0/24
      egressGroups:
      - nodeSelector:
	  matchLabels:
	    labelKey: labelVal
	interface: eth2
	maxGatewayNodes: 2
      - nodeSelector:
	  matchLabels:
	    labelKey2: labelVal2
	interface: eth2
	maxGatewayNodes: 2

Enterprise-feature: Egress Gateway HA
Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>