egressmap: add support for healthy gateways
This commit updates ApplyEgressPolicy to take an additional parameter,
healthyGatewayIPs, which will be used to skip removing egress gateway CT
entries belonging to gateway nodes which are not active anymore, but are
still healthy.

This is to avoid breaking existing connections in case the list of
gateways for a given policy gets shuffled.

Enterprise-feature: Egress Gateway HA
Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>