bpf: Implement SRv6 decapsulation on ingress
For any IPv6 packet received on ingress of the native devices, we check
if the destination IP address is a known SID. If it is, we tail call
into a program to perform the actual decapsulation. That simply consist
in removing the outer IPv6 header and, if the inner header is IPv4,
changing the protocol on the skb. After that, the packet is sent ot the
stack.

Signed-off-by: Paul Chaignon <paul@cilium.io>