srv6: Expose VRFs and egress policies for BGP manager
Depending on the configuration, the BGP manager will be in charge of
creating the CiliumSRv6EgressPolicies based on the routes received via
BGP. To that end, it needs to know what VRF exists (to map BGP
announcements to the different VRFs) and what egress policies are
already loaded (to create/delete the missing/outdated ones).

The SRv6 manager has ownership of those data structures, so it will
expose it to the BGP manager via a set of getter functions. In addition,
only fields required by the BGP manager are exposed. The BGP manager for
example doesn't need to know about the VRF rules.

Signed-off-by: Paul Chaignon <paul@cilium.io>