bpf: Implement Segment Routing Header support
Until now, we have only supported the reduced encapsulation format for
SRv6. That was possible because we also only support a single segment
for now so we can always use the reduced encapsulation format.

However, some routers (e.g., FRR) don't support reduced encapsulation
yet. Thus, we should also have a fallback to include a full Segment
Routing header when encapsulating and decapsulate accordingly.

The logic is fairly straighforward: we just need to include offset to
add and remove a new header, as well as populate it.

When decapsulating, we need to jump to the code parsing the IP header
once the SRH is parsed. We can't do that by jumping to the beginning of
the switch because the verifier would complain that we have an infinite
loop (not yet smart enough to understand it can't happen). Instead, we
include a switch in the switch :'(

Note with this commit we still don't support both modes at the same
time: we expect the same outer header on ingress that we sent on egress.
So when reduced encapsulation is disabled, we expect ingressing SRv6
packets to have an SR header.

Signed-off-by: Paul Chaignon <paul@cilium.io>