bpf: Fix typo in host firewall tail call
[ upstream commit 56bbbedfb8adf49ee95dca2ce336f7735104ccd4 ]

There's a typo in the host firewall tail call that can lead to us
executing the IPv6 code for an IPv4 packet when enforcing host policies
for a packet coming from a local pod, when both IPv4, IPv6, and endpoint
routes are all enabled. The typo was found while working on splitting
bpf_host into more BPF programs.

The next commit explains why the end-to-end tests didn't uncover this
bug and improves those tests.

Fixes: 5b05cc92 ("bpf: Split handle_lxc_traffic with tail calls")
Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>