test: More fine-grained host policies
[ upstream commit 61f15b36a01431869d5b7654714c08bc40f3d2eb ]

The previous commit fixed a bug in the host firewall where IPv4 traffic
from local pods would be incorrectly processed as IPv6 traffic. That
path is covered by several end-to-end tests in our CI but none of them
failed.

In those tests, we perform a TCP request on port 80 and a UDP request
on port 69; the first is allowed by policies, while the second isn't and
should be dropped. The requests are allowed and denied as expected on
the bogus path. So let's take a look at how that's possible.

We don't have any policy verdict for the TCP request, but we have one
for the reply. So the TCP SYN packet skipped policy enforcement
altogether.

    $ ks exec cilium-cgg9s -- cilium monitor -t policy-verdict
    Policy verdict log: flow 0xd9cfbd98 local EP ID 1630, remote ID 3812, proto 6, egress, action allow, match L3-Only, 192.168.56.11:80 -> 10.0.0.240:39134 tcp SYN, ACK

Because of the bug, the IPv4 SYN packet went through
ipv6_host_policy_ingress. The first step of that function was to lookup
the destination security ID using the destination IP address. If the
destination security ID is *not* HOST_ID, then host policy enforcement is
skipped. Unsurprisingly, looking up an IPv4 address in the ipcache with
an IPv6 prefix returns nothing and the security ID is set to WORLD_ID.
Host policies are thus skipped and the SYN packet goes through.

Now, why was the SYN+ACK allowed by policies? The above policy verdict
says it comes from an L3 rule. Looking at the policies, we do have an L3
egress rule to allow pods not involved in the test to communicate with
the host (on egress, only hostns -> zgroup=testserver requests are
expected):

    - toEndpoints:
      - matchExpressions:
        - key: zgroup
          operator: NotIn
          values: [testServer]

Unfortunately, that ends up matching the return traffic of the
zgroup=testclient -> hostns connection.

But then, why wasn't the UDP request allowed through as well? That would
have failed the test and we would have caught this bug. If we look at
policy verdicts for the UDP request, we do find the same sort of allow:

    $ ks exec cilium-cgg9s -- cilium monitor -t policy-verdict
    Policy verdict log: flow 0x0 local EP ID 1630, remote ID 3812, proto 17, egress, action allow, match L3-Only, 10.0.2.15:69 -> 10.0.0.240:54715 udp

So the request isn't blocked by policies. If we look at drops:

    $ ks exec cilium-cgg9s -- cilium monitor -t drop
    xx drop (Invalid packet) flow 0x0 to endpoint 0, , identity unknown->unknown: 10.0.0.240:49043 -> 10.0.2.15:69 udp

The most likely reason for an Invalid packet drop is that we received a
packet that is too small (e.g., we're trying to read the IPv4 header but
the packet is smaller than the minimum IPv4 length). So in this case, we
received an IPv4+UDP packet that is smaller than the minimal IPv6 length
and thus dropped it. Hence, the UDP connection fails, as expected by our
test.

This commit improves the test by making sure the L3 rule applies only to
pods not involved in the test, neither as servers nor as clients.

Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>