Add support for disabling ENI PD at node level
[ upstream commit 24a7e412b1196e829d76c22ad44613d56937d58b ]

Currently prefix delegation (PD) can only be enabled cluster wide. On
nodes with low pod density, PD can be wasteful in terms of number of
unused IP addresses allocated to the node. This commit adds a new field
to the ENI spec on ciliumnode CRD to opt out of PD at a node level.
Other IPAM settings like pre and min allocate are already defined on the
CRD.

[ Backport notes: Conflict: Jumping
    CustomResourceDefinitionSchemaVersion from 1.25.3 instead of 1.25.4
    in pkg/k8s/apis/cilium.io/v2/register.go. ]

Signed-off-by: Hemanth Malla <hemanth.malla@datadoghq.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>