option: fix config map flags validation
[ upstream commit 98280fad7ca49704428e022950f2bc58131fdcbd ]

Through the ConfigMap, all config values are loaded as strings and then
converted into the appropriate types. The conversion is carried out by
Viper that does not report any kind of error if something goes wrong.

To catch early any kind of conversion issue, we should use the same
conversion APIs that Viper is using internally, so we check each option
using the appropriate conversion function from the cast package.

Related to #20282

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>