daemon: Deprecate KPR=probe
[ upstream commit 7463d9f12ba643d707f95a587ea03107ae46f0cb ]

The option is a never ending source of confusion. Also, after changing
system configuration, the "probe" option could unexpectedly enable some
KPR features which might be not expected by users.

Deprecate the "probe" option for now. In v1.13 we will make the option
to switch to "disabled". Unfortunately, we cannot just remove the
option, as it was used as a default in the Helm charts.

Signed-off-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>