bgp: fixed PodCIDR announcement being overwritten by SVC announcement
[ upstream commit 46acab7edaeb42ab0f1cb1986ff26a8f6184fcb7 ]

This work fixes a bug, where the PodCIDR announcements would disappear
as soon as another announcement like a loadbalancer was added. The root
cause turned out the be that we created the PodCIDR announcements
directly in the BGP sessions, however the session announcements are
completely overwritten by the BGPController which maintains its own
source of truth. So we now add the PodCIDR announcements to the
BGPController and ask it to sync. This makes it so we play nice with
the other announcement types.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>