datapath: add file and line info to drop notifications
[ upstream commit f3f1f28c686f111a5dbf9d7ac0c1978461c7e35c ]

Add source file and line information to the output of `cilium monitor -t drop`.
The information is passed to user-space in the new `line` (u16) and `file` (u8)
fields of the struct drop_notify structure.  In order to pass this information
over through a tail call the 24 upper bits of the meta[4] field were used.
Previously meta[4] kept only the exit code which is a small number (XDP or TC
return codes), and should fit in the dedicated 8 bits:

        8           8             16
  | exit code | file number | line number |

The file number is defined in the loader.sourceNameToId map which maps all C
files from the bpf/ directory to ids.  The map is to be maintained manually
(in the initial version of this patch it was generated by go generate, but
this could lead to inconsistencies if agent and client were built separately).
The consistency of this map is checked during build, so if one adds a new C
file to the bpf/ directory and doesn't update the map, then the build fails
(see make check-sources).

In order to send this info we need to gather it in the first place.  To do this
with minimal changes to the codebase the send_drop_notify{,_error} functions
were redefined as macros:

    send_drop_notify(...)       -> _send_drop_notify(__MAGIC_FILE__, __LINE__, ...)
    send_drop_notify_error(...) -> _send_drop_notify_error(__MAGIC_FILE__, __LINE__, ...)

The __LINE__ expands to the line number.  Be aware though that if a macro call
expands to several lines, like here

    581       if (IS_ERR(ret))
    582               return send_drop_notify(ctx, SECLABEL, dst_id, 0, ret,
    583                                       CTX_ACT_DROP, METRIC_EGRESS);

then clang translates the __LINE__ to a slightly wrong number—the last line the
macro call occupies, 583 in this case. gcc would translate it to 582.

The __MAGIC_FILE__ constant is redefined according to the loader.sourceNameToId
mapping by the loader when we compile a program.  (If the __MAGIC_FILE__ macro
is not defined during the compilation it defaults to zero.)

With all these changes applied, one now can see corresponding filenames and
line numbers in the output of `cilium monitor -t drop`:

    xx drop (Invalid source ip) flow 0x0 to endpoint 0, file bpf_lxc.c line 695, ...
    xx drop (Stale or unroutable IP) flow 0x15a7e11f to endpoint 0, file bpf_host.c line 351, ...

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>
Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>