bpf: Handle tuple collisions on services for inactive backends
[ upstream commit 8600cdd46b94bfc24d13d298c77710c372bdd8cc ]

Yusuke reports:

  Graceful termination does not seem to work as expected in a particular
  case where a client accidentally picks a stale source port. CT_SERVICE
  entries remain for six hours because FIN/ACK never comes to intermediate
  nodes. So clients are likely to hit the same source ports with the DSR
  mode.

The work in #19451 / #19800 partially resolves this case: Cilium selects a
new backend if the ct_service entry is closing and CT_SERVICE_CLOSE_REBALANCE(30)
seconds have passed since the last TX. If bpf-ct-timeout-service-tcp-grace is
set to 0, Cilium always selects a new backend when it receives an SYN packet
that has the same old tuple. Downside being that this reduction would ignore
TIME_WAIT state.

The case when backends are inactive, they are either terminating (and can
therefore disappear at any point in time), quarantied (e.g. because connectivity
may be flaky), or placed into maintenance mode (where we expect to drain traffic)
and thus for this corner case we might choose to pick a new, active backend.

After the fix, Yusuke mentioned:

  I could confirm this PR works as expected in our environment. Cilium
  selects new backends for new connections after the Envoy becomes the
  terminating state, and the terminating Envoy could properly drain active
  connections. There were no communication disruptions during the rolling
  restart.

Fixes: #20382
Reported-by: Yusuke Suzuki <yusuke-suzuki@cybozu.co.jp>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Tested-by: Yusuke Suzuki <yusuke-suzuki@cybozu.co.jp>
Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>