docs: add missing flag to rollout pods in egress gateway
[ upstream commit 68e4ad6aaf869d90aaa36273f7fa7c001942bb0b ]

The `helm upgrade` command depicted in the documentation for the Egress
Gateway feature does not report the use of the flag
`--rollOutCiliumPods=true`.  Without this flag, the operator and the
agent pods will not be restarted, thus the new values in the ConfigMap
will not be loaded.

To avoid failures while following the next steps of the guide, the
commit adds the above flag is added.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>
Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>