helm: Set KPR default to "disabled" for >= 1.12
[ upstream commit 74d719cb0b59ecd701a264dccc9b892c451e69e9 ]

The commit 7463d9f12 ("daemon: Deprecate KPR=probe") deprecated the
KPR=probe option. However, it missed to explicitly set the default for
>= 1.12 installations.

The result of it is that the "probe" is used as a default for a new
installation. This is because the following condition evaluates to true
when ".Values.upgradeCompatibility" is not set:

    {{- /* Default values when 1.9 was initially deployed */ -}}
    {{- if semverCompare ">=1.9" (default "1.9" .Values.upgradeCompatibility) -}}
      {{- $defaultKubeProxyReplacement = "probe" -}}
    {{- end -}}

NB: I find this very confusing and error prone. The following makes more
sense:

- the default ".Values.upgradeCompatibility" should be the latest
  version
- "==" instead of ">=" should be used

Otherwise, all "semverCompare" if-clauses we have will evaluate to true.

Fixes: 7463d9f12 ("daemon: Deprecate KPR=probe")
Signed-off-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>