iptables: handle case where kernel IPv6 support is disabled
[ upstream commit 03b89ec25edc538f9627b3e82a9e698ce7bba1a0 ]

Currently, even if the kernel IPv6 support is disabled (i.e. the
ipv6.disable=1 kernel parameter is set), the agent will try to find or
load the ip6tables kernel modules, and if successful it will assume
ip6tables is supported.

This will result in a fatal error since all ip6tables commands will
fail with "Address family not supported".

This commit fixes this by setting the haveIp6tables IptablesManager's
property to false in case the kernel does not support IPv6.

Fixes: d812b925de ("iptables: don't ignore errors")
Fixes: #18513
Fixes: #20672
Signed-off-by: Gilberto Bertin <jibi@cilium.io>
Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>