tools/customvet: consider $GO environment variable
[ upstream commit 920094b35dc22bbd92fb9c7738ceb8d9597f5277 ]

For backports it's usually convenient to be able to set the $GO
environment variable to the version used in the particular stable
branch, e.g. on v1.10 we'd set GO=go1.16.15. However, that setting is
currently not considered in tools/customvet. This can lead to unexpected
failures such as the followin when running `GO=go1.16.15 make precheck`
on v1.10 branch:

    contrib/scripts/custom-vet-check.sh
    main: internal error: package "sync" without types was imported from "github.com/cilium/cilium/pkg/cleanup"
    exit status 1

Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>