fix k8s latency metrics label cardinality
[ upstream commit 4ad5e94f0c0c99fee6ebcaf4bb2ab3933a17c7d0 ]

The functionality provided by the finalURLTemplate is still used by
Cilium to track the request latency for requests performed to
Kubernetes. Until this is reverted upstream we will maintain this fork
with this change so that we can still used the functionality provided by
finalURLTemplate.

This functionality prevents the explosion of label cardinality in
prometheus metrics since it aggregates in a way that common URLs
requests will be reported as being the same.

Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>