iptables: skip NOTRACK rules deletion
[ upstream commit 4bd2478db37a6859c57372dcc97ac43922d26e90 ]

The Iptables version shipped with Cilium (1.8.4) does not fully support
the iptables-legacy - nft translation of NOTRACK rules, which in turn
causes the agent to crash whenever it tries to delete any old/backup
ruleset.

This commit introduces a workaround for this by ignoring the `-j
NOTRACK` and `-j CT --notrack` rules when deleting all rules belonging
to an old/backup ruleset.

This is not an issue (although it's not optimal) as eventually the agent
will just flush and remove all the Cilium managed Iptables chains.

Fixes: #20714
Signed-off-by: Gilberto Bertin <jibi@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>