srv6-l3vpn-ingress: provide IPv6Allocator to SRv6 Manager
This commit updates the daemon init function to optionally pass the
IPv6Allocator managed by IPAM to the SRv6 Manager.

Due to the order in which subsystems are initialized in the Daemon, a
setter is used to provide the SRv6 Manager the IPAM allocator, rather
then providing it during construction.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>