srv6,Mananger: handle agent lifecycle events
This comit refactors the code to properly handle agent restart events.

We can remove the `runReconciliationAfterK8sSync` due to the fact we will
obtain events after the K8s sync occured and the initial events for existing
CRDs are dispatched.

We can then tie reconciliation events to the setting of our dependencies, so we
attempt to reconcile when the BGP manager and SID Allocator are set.

Lastly, we update the SID removal function to remove any SIDs in the SID map
which we do not have allocations for. This will clean up any stale SIDs on an
agent restart.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>