test/srv6: Fixup SID map manually
We need to manually teach k8s1's datapath about the expected SID for
End.DT{4,6} on ingress. This is because, when running SRv6 without
the BGP integration, we don't populate the SID map. It needs a proper
fix longer term (or we drop CRDs) but for now let's just fixup the
test since the BGP integration is what users are expected to run.

Note we can hardcode k8s1 here because we know the client pod always
runs on k8s1 (see manifest.). The VRF ID is hardcoded as well, for
the same reason.

Signed-off-by: Paul Chaignon <paul@cilium.io>