clustermesh: import phantom services into ServiceCache
When the service coming from the remote kvstore is marked as "phantom",
import the service into ServiceCache. The ID of such service will
contain the remote cluster name, to prevent collisions if declared as
phantom by multiple clusters.

This commit introduces a wrapper around the OSS ServiceCache, which
overrides the cluster service merging logic to support both standard
global services and phantom services. The actual implementation is
swapped through an Invoke function executed as part of the clustermesh
enterprise cell.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>