operator/dnsresolver: Add CIDRGroup reconciler
A CIDRGroup reconciler is associated to a single FQDN group and aims to
reconcile the status of the related CiliumCIDRGroup object. It accepts a
name and a UID to add an ownerReference to the managed CiliumCIDRGroup
object, in order to delegate the GC of the CiliumCIDRGroup to the k8s
apiserver.

A reconciler listens for events from each FQDN resolver whose FQDN is
listed in the associated IsovalentFQDNGroup. It also contains a cache,
in order to avoid calling the apiserver if no actual update to the
CIDRGroup should be performed.

Co-authored-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>