enterprise: Bootstrap daemon and operator
Add enterprise daemon and operator as distinct binaries from their OSS
counterparts.

The enteprise binaries rely on separate cobra.Command(s) and hives to
extend the OSS features limiting as much as possible the changes to the
shared part of the codebase.

Also, both enterprise agent and operator relies on a separate go.mod.
This is done to avoid conflicts when additional packages are imported by
enterprise-only features.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>