clustermesh: add phantom service tests to cilium-cli
This commit leverages the hooks framework to extend the cilium CLI
connectivity tests with additional scenarios to verify that phantom
services work properly. A setup function is furthermore registered
to create the phantom service in one of the clusters, and configure
fake LoadBalancer IP addresses (selected among the ones reserved
for documentation purposes); a dummy LoadBalancer class is also set
to prevent real implementations from reconciling this service.
To prevent race conditions, the setup function waits until the service
information has been propagated to the agents running on the nodes
hosting the client pods. A finalizer is registered to remove the
service when the test completes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>