Add false positive handling using VEX documents
This PR adds a [VEX document][0], which is a standardized way of adding contextual
information (such as false positive annotations) to the results generated by
security tooling.

We use the VEX document in conjunction with `vexctl` and our initial scanner
results to create a filtered set of vulnerabilities that excludes all false
positives that we have already reviewed. Adding more false positives is simply
a matter of editing the VEX document in the repository.

This presents the reviewer of our container scans with a much cleaner report
to assess.

Also switch from `on: pull_request` to `on: workflow_dispatch`, which I set
incorrectly when I created the workflow.

[0]: https://cyclonedx.org/capabilities/vex/

Signed-off-by: Feroz Salam <feroz.salam@isovalent.com>