cilium-cli(hooks): Add hook to setup and validate context
This commit intents to pass the context from connectivity test to enterprise
tests. This is done by creating a new hook SetupAndValidate() taking a context
as parameter.  Functions like detectFeatures() and
RetrieveExternalCiliumDNSProxyPods() are now called in this new hook and no
need to create a new context anymore. It allows to guarantee continuity during
the test execution if the context is cancelled for example.

Signed-off-by: Donia Chaiehloudj <donia.cld@isovalent.com>