ci: replace GHA action Sibz/github-status-action
This commit replaces the GitHub Action `Sibz/github-status-action` to
set the commit status with `myrotvorets/set-commit-status-action`.

Sibz/github-status-action is only used in the Helm Chart Push and builds
on top of node12 which is deprecated in GHA.

`The following actions uses node12 which is deprecated and will be
forced to run on node16:
Sibz/github-status-action@650dd1a882a76dbbbc4576fb5974b8d22f29847f. For
more info:
https://github.blog/changelog/2023-06-13-github-actions-all-actions-will-run-on-node16-instead-of-node12-by-default/`

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>