bgp/speaker: use Resource[T] instead of NodeChains
As part of a refactoring effort to remove the CiliumNodeChain, remove
the dependency of the MetalLBSpeaker on it, by directly using the
LocalCiliumNodeResource.

Note that the speaker always discarded events not pertaining to the
local Node, hence no functional change results from the fact that the
CiliumNodeChain proagated events from all cluster CiliumNodes.

As an added benefit, the CiliumNode events now don't stop coming once we
connect to KVStore, since the resource doesn't stop.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>