egressgw HA: rename copied packages
Give the Egress Gateway HA packages distinct names from the OSS feature. This
simplifies code navigation and also helps disambiguate between the two features
in log output. For example, hive logs start and stop constructors in the form

    package.Function

Without renaming the packages it is extremely easy to mistake one feature for
the other.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>