policy: Make MapState an interface
`MapState` has many methods, some of which are public.
In order to keep state to accomplish creating a cohesive
map state that can keep state besides what will be put into
an ebpf map, `MapState` can no longer be a type declaration
for a golang map, but must be a struct that keeps state.

The endpoint package must also be refactored to use `MapState`
as an interface rather than a instantiated type of a map.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>
Signed-off-by: David Bimmler <david.bimmler@isovalent.com>