srv6: Implement some helper functions for SIDInfo
We'll change the type of VRF.AllocatedSID to SIDInfo. To do that, we
need to implement SIDInfo to make DeepCopy in SRv6Manager.GetAllVRFs
work.

We also need the way to compare if the SID and Behavior are the same as
another SIDInfo, so that we can compare the set of SIDInfo. This is
useful for reconciliation.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>