ent/cli: check: add support for IsovalentEgressGatewayPolicies
allow enterprise tests to deploy IsovalentEgressGatewayPolicies with the
WithIsovalentEgressGatewayPolicy method.

This is implemented by decorating the upstream ConnectivityTest object
as a EnterpriseConnectivityTest, which contains the enterprise k8s
client used to access these resources.

With an EnterpriseConnectivityTest object we can then create a new
enterprise test with NewEnterpriseTest, which will take care of setting
up the iegp policies.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>