bpf,srv6: move h.encap to bpf_lxc.c
This commit moves the srv6 h.encap functionality to bpf_lxc.c

In this change we determine if the egress traffic belongs to a VRF early
in the pod egress data path.

If it does we use ctx->cb to both signal this and stash the VRF ID for
later lookup.

The lookup and subsequent h.encap is then performed in
"handle_ipv{4,6}_from_lxc" functions, after policy lookup occurs.

As of today, we do not support l7 policy, thus we do not even try to
redirect to proxy for VRF traffic, this may change in the future.

This commit makes the bpf_host.c handling of h.encap redundant and a
follow up commit will remove this.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>