test: Update Go tests for renamed cilium-dbg
This commit changes most of the existing shell commands for the local
Cilium API over to 'cilium-dbg', but skips the core helpers. The reason
for this is that the core helpers are used by the Updates test which
runs the commands against an older version of Cilium. As such, we cannot
replace those commands with 'cilium-dbg' until the next development
version when we start updating from a Cilium copy that has the
'cilium-dbg' binary inside.

Signed-off-by: Joe Stringer <joe@cilium.io>