Rename enterprise/cilium to enterprise/cilium-dbg
Due to the recent change in the upstream (cilium/cilium#28085), cilium
(node-local command) is renamed to cilium-dbg. This broke some
enterprise-specific pieces such as Makefile target and Go imports.

Also, when we do `make -C Documentation update-cmdref`, it generates
command reference documents prefixed with `cilium_` and it conflicts
with the upstream because upstream docuements are prefixed with
`cilium-dbg_`.

To fix these issues, rename enterprise/cilium directory to
enterprise/cilium-dbg. Namely, renaming enterprise (local) cilium
command to cilium-dbg as well.

This commit also updates the references to enterprise (local) cilium
command and regenerate the cmdref documents.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>