egressgwha: move gateway selection logic to the operator
This commit moves the egress gateway manager logic that deals with
gateway selection, mostly figuring out which nodes are active/healthy
gateways for a given policy, from the agent to the operator.

Specifically the operator will now consume CiliumNode and
IsovalentEgressGatewayPolicy k8s events, will perform healthchecking of the
gateway nodes, and will publish the status of the active/healthy
gateways into the newly introduced groupStatuses field.

The agent will then consume this new field, and together with the
IsovalentEgressGatewayPolicy and CiliumEndpoint k8s events it will run the
usual reconciliation logic to update the state of the bpf maps and the
eventual list of IP rules/routes.

It's worth nothing that when the groupStatuses.observedGeneration is
older than the policy one, which means it has been updated by the user
but the operator has not finished yet updating the groupStatuses, the
policy is simply ignored, as otherwise the manager would end up applying
a configuration of gateways from groupStatuses which is out of sync with
the policy's desired state.

This is needed as with the previous approach each agent was running its
own checks to determine the list of gateways for a given policy, and
there were no guarantees that all of them would agree on the same list.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>
Signed-off-by: Lorenz Bauer <lmb@isovalent.com>
Signed-off-by: James Laverack <james@isovalent.com>