add endpointslice controller for IsovalentMeshEndpoints
In order for transit gateway to make IME part of ServiceBackends, we
will use K8s EndpointSlices for this purpose.

To avoid re-writing the same logic already available in K8s upstream, we
will be using a private fork, with a couple of adjustments so that it
can be used for our use cases, in Cilium for the CiliumMesh feature.

Signed-off-by: André Martins <andre@isovalent.com>