srv6: Remove SRv6 Manager from Daemon's dependency
Remove SRv6 Manager from Daemon's dependency and use DaemonPromise with
SRv6 Manager. We used to obtain the reference to the IPAM allocator
within newDaemon, but now we need to get the Daemon first and get IPv6
IPAM Allocator from it. However, the Daemon doesn't expose the Allocator
to the external module. That's why we needed to introduce
GetIPv6IPAMAllocator() method.

Also, mocking out the daemon in the test is hard, so we introduced a
private daemon interface which only implements GetIPv6IPAMAllocator()
and wrap Damon's promise with the promise returns the daemon interface.
This allows us to introduce fakeDaemon in the SRv6 Manager test.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>