enterprise: Add multinetwork operator cell
This commit adds a new operator cell to automatically create the
"default" network on startup. This is required for any pod which does
not attach to any special network. If the "default" network is not
present, all such pods would be stuck in ContainerCreating state until a
"default" network is created by the user.

In addition, this commit also adds Helm values for multi-network.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>