ci: Add multi-network enterprise workflow
This tests the Cilium-native multi-network (or multi-homing) feature in
enterprise by running connectivity tests, verifying that CEPs got
created for multi-homed pods and that these have their routes set up for
the secondary network.

Because a secondary docker network is needed for the tests, but that
cannot be created when using the helm/kind-action action, a LVH VM is
used in which the kind cluster and the secondary docker network are
created.

The workflow is triggered via `/ci-multi-network` or `/test`.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>