egressgatewayha: Use UID to identify CiliumEndpoints in epDataStore
To address a potential race condition demonstrated by a unit test
added as part of this change, this changes internal type of the endpointID
used to identify CiliumEndpoints in the EGW Manager's epDataStore
from NamespacedName to UID.

The race fixed by this change can be triggered during statefulset pod
restart/migration to a different node, when a new endpoint may co-exist
with the to-be-deleted endpoint with the same NamespacedName for a short
period of time. To not break EGW Manager's endpoint event handling, these two
need to have unique endpointID, which can be satisfied by relying on the UID.

The drawback of using UID is incompatibility with CiliumEndpointSlices,
which is however not an issue for the EGW, as it is already incompatible.
In the future the EGW will probably move away from relying on
CiliumEndpoints altogether, which should remove this limitation.