cilium-cli: Fix Tetragon EE tracing policies
In order to get the correct Tetragon EE tracing policies we need to use
github.com/isovalent/hubble-fgs/pkg/k8s instead of
github.com/cilium/tetragon/pkg/k8s.

This replace requires the use of a private repo. One way was to update
all CI jobs to use GOPRIVATE etc. We choose not to take this approach as
we would like to minimize the changes in upstream files.

Adding a vendor directory inside enterprise/cilium-cli was not a
reasonable approach due to
https://github.com/isovalent/cilium/pull/1188#pullrequestreview-1441864267.

For this reason, we keep a copy of
https://github.com/isovalent/hubble-fgs/tree/master/pkg/k8s inside
enterprise/cilium-cli/tetragon. This is managed by normal go modules.
This approach does not require any changed to any files outside of
enterprise/cilium-cli.

FIXES: https://github.com/isovalent/hubble-fgs/issues/3184

Signed-off-by: Anastasios Papagiannis <tasos.papagiannnis@gmail.com>