ci: wait for rollout to complete in enterprise kind workflow
Wait for the rollout of ds/cilium and ds/cilium-dnsproxy before
continuing with the enterprise tests. Also wait for ds/cilium-dnsproxy
instead of ds/cilium after installing FQDN proxy. However, we currently
need to disable this check because the agent is not yet started with the
corresponding flag. The rollout wait will be enabled in a follow-up
change.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>