Override CILIUM_CLI_VERSION
With cilium/cilium#29237 and cilium/cilium#29338 merged upstream, we can
now override and use the enterprise version of cilium-cli for running
tests. The goal is to backport this to v1.14-ce branch so that we can
use the enterprise cilium-cli released from the main-ce branch for
running tests in stable branches starting from v1.14-ce.

Ref: https://isovalent.slack.com/archives/C05188W95GT/p1700144241292359

Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>