hubble-flow-aggregation: Fix nil aggregation plugin being returned to hubble-flow-export
Plugins are initialized, then their dependencies are injected, and much
later OnServerInit() is called. Because dependencies are added
immediately after initializing the plugins, export was calling
GetFlowAggregator before the aggregator was initialized in OnServerInit,
causing a nil aggregator to be returned to the export plugin.

I want the logger to be available from the beginning, so instead of
initializing the aggregator within OnServerInit, initialize it in the
New function of the plugin.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>