mixed routing: add configuration flag and validation
Introduce a new configuration object to register the flag inherent mixed
routing mode support. The flag takes as value the fallback mode leveraged
in case of a mismatch between the routing mode configured in the source
and destination node. Currently, the only supported option is tunnel
routing, although native routing support is foreseen for the future.
Extra validation logic is implemented to reject known incompatible
configurations, mainly with respect to IPAM mode and NodePort DSR support.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>