mixed routing: enable tunneling when fallback routing mode is tunnel
When the primary routing mode is configured to native routing, and the
fallback one is tunneling, we need to enable the configuration of the
tunnel device, and the compilation of the corresponding BPF logic.

Let's extend the mixed routing mode cell to inject the appropriate
parameters into hive, in order to instruct the corresponding upstream
modules to do so.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>