mixed routing: add utilities for supported routing modes advertisement
In order to enable mixed routing mode, each node needs to start
advertising the supported routing modes, including the primary one
and, if different, the configured fallback. This information will
be conveyed through a dedicated annotation attached to the CiliumNode
resource (and the corresponding kvstore representation), such as:

  routing.isovalent.com/supported: native,tunnel/geneve

Which means that the preferred routing mode for this node would be
native, with supported fallback to tunneling mode (using the Geneve
protocol). The tunnel protocol is included in order to detect possible
incompatibilities (i.e., two nodes both support tunneling, but with
different protocols). As well as to be more future proof if mixed tunnel
protocol support will be introduced in the future.

This commit introduces the type definitions and the utility functions
to serialize the routing modes into the annotation string, and parse
it back into a slice of modes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>