mixed routing: configure supported routing modes annotation
This commit introduces the manager implementing mixed routing mode
support. At the moment, it takes care of initializing the list of
routing modes supported by the local node, and attaching it to the
local node object through the dedicated annotation. Eventually,
they will propagate to the corresponding CiliumNode resource,
as well as to the respective kvstore representation.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>