enterprise/api: Add new GET /network/attachment endpoint
This adds a new endpoint to the Cilium API which allows the CNI to
obtain the network attachments of a particular pod. The API responds
with a list of network attachments that the user has requested for the
pod, which means the CNI is expected to set up one interface for each.

This commit only contains the model changes and stub handlers. The API
will be implemented in a subsequent commit.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>