mixed routing: introduce node manager
Introduce the node manager, responsible for handling clustermesh
related node upsertion and deletion events. More specifically, it
implements the logic to determine whether encapsulation needs to
be configured for the given node (to be used to tune the behavior
of the OSS node manager logic.

Moreover, it takes care of emitting a synthetic deletion event if
the preferred routing mode towards that node changes. This ensures
that all required datapath changes (e.g., modification of the
node routes) are transparently executed by the OSS logic, without
requiring further extensions to handle this special case.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>