mixed routing: hook node manager into place
Let's hook the node manager introduced in the previous commit into
the OSS logic, leveraging appropriate injection functions. In
particular, we wrap the node observer leveraged by the clustermesh
subsystem to observe all node update/delete events. Additionally,
we override the function used by the OSS logic to determine for
each node whether encapsulation shall be configured or not, based
on the matching routing mode.

The extra node manager logic is hooked into place only when a fallback
routing mode is enabled, and it is different from the primary routing
mode (as otherwise the primary would be always selected anyway). Yet,
in that case, we still wrap the node observer with a lightweight
version in charge of only logging an error message if the advertised
routing mode is incompatible with the locally configured one.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>