mixed routing: filter ipset entries creation based on routing mode
Normally, Cilium inserts an ipset entry for each NodeInternalIP address
when configured in native routing mode if iptables masquerading is
enabled, to skip masquerading for pod to node traffic. Differently,
masquerading must be performed when running in tunnel mode, otherwise
packets may be dropped, and return traffic would not work.

This commit extends the mixed routing node manager to additionally
filter out insertions into the ipset for the InternalIP addresses
of nodes configured in tunneling mode, in case the primary routing
mode is set to native (which would normally cause the configuration
of the ipset entries). Specifically, this is achieved overriding the
IPSetManager leveraged by the OSS logic with a wrapper, which takes
into account the exclusion list configured when new node upsertions
and deletions are observed. ipset deletions are assumed to be no-op
if the entry is not already present, and are always propagated.
Additionally, possibly existing stale ipset entries are deleted
if the corresponding insertion was filtered out.

The logic herein introduced works under the assumption that the
fallback routing mode is tunnel (i.e., the only supported configuration
at the moment). Hence, it filters out the ipset entries associated
with nodes towards which the preferred routing mode is tunnel.
To support native routing fallback, instead, we should force the
creation of the ipset, and manually insert the appropriate entries.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>