mixed routing: introduce endpoint manager
Introduces the mixed routing endpoint manager, responsible for keeping
tunnel endpoint to routing mode mappings and handling clustermesh-related
endpoint upsertion and deletion events. Specifically, it implements a
wrapper around the ipcache to buffer the upsertion of entries associated
with host IPs for which we haven't yet seen the corresponding node event,
as we don't know their associated routing modes. This is key to avoid
needing to additionally attach routing mode information to all endpoints.

Once a given mapping is configured, all buffered endpoint are then
automatically inserted into the underlying ipcache subsystem. Similarly,
upon routing mode change, a synthetic deletion event, immediately
followed by a new upsertion, is emitted for all entries associated with
that tunnel endpoint, to make sure that the change is eventually
propagated to the ipcache BPF map (a simple upsertion may not be
sufficient, as it could be short-circuited by the ipcache subsystem if
none of other parameters changed).

Besides endpoint entries associated with a non-zero tunnel endpoint, we
do also care about those corresponding to the Node{Internal,External}IP
addresses of the nodes themselves (as required to tune BPF masquerading).
Yet, we do not need any buffering logic in that case, as they get
inserted synchronously by the node upsertion event (as well as they are
already removed and inserted back following a routing mode change).

Buffering does not introduce concerns about the detection of the initial
synchronization of the ipcache (which is required to prevent disrupting
cross-cluster connections on agent restart), because the wait function
already accounts for nodes synchronization in addition to IPs and
identities. And upon node event reception, we synchronously propagate
all the corresponding buffered ipcache entries.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>