mixed routing: introduce ipcache map value mutator
Extend the mixed routing endpoint manager with the logic to appropriately
mutate the skip_tunnel flag part of the ipcache map value depending on whether
encapsulation is the preferred routing mode towards that specific endpoint.
In particular, endpoints are matched based on the tunnel endpoint value or,
if unset, the prefix itself (only when it represents a single IP address).
This latter case covers the entries corresponding to the NodeInternalIP and
NodeExternalIP addresses of the nodes themselves, which are required to tune
BPF masquerading. As a final fallback, we default to the local routing mode
for all other non-matching entries.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>