mixed routing: add utilities for supported routing modes matching
Introduce an helper method to return the first matching routing mode
present both in the local and in the remote list, or an error in case
no match is found.

This method is guaranteed to lead to a symmetric selection of the
routing mode as long as every node is configured with a primary mode
and exactly the same secondary mode, possibly matching the primary
one and omitted (as in the current implementation). For instance,
given:

  local=native,tunnel/geneve and remote=tunnel/geneve,

then:

  local.match(remote) == remote.match(local) == tunnel/geneve.

Differently, it could lead to different decisions on different nodes if
associated with multiple unsorted secondary modes. For instance, given:

  local=native,tunnel/geneve,tunnel/vxlan and remote=tunnel/vxlan,native

then:

  local.match(remote) == native, remote.match(local) == tunnel/vxlan.

To support this use-case ensuring symmetric routing, we would need to
additionally introduce a priority order in case multiple matches are
found (e.g., native > tunnel/geneve > tunnel/vxlan). But this is out
of scope for the moment.

Additionally, let's also introduce an helper to retrieve the primary
routing mode from a list of routing modes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>