mixed routing: buffered endpoints observability
Let's introduce the possibility of observing the number of endpoints
buffered waiting for the hosting node information, to allow
troubleshooting possible issues. Specifically, we introduce a dedicated
gauge metric, which shall always converge to zero after the initial
synchronization phase. In addition, the list of prefixes is periodically
scanned and a log warning emitted for entries buffered for more than
one minute, to provide more detailed information about them.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>